backend/pyproject.toml
[build-system]
requires = ["setuptools>=75.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cropfolio"
version = "0.1.0"
description = "Portfolio Theory for Climate-Resilient Farming in Myanmar"
requires-python = ">=3.10"
dependencies = [
    "fastapi==0.115.6",
    "uvicorn[standard]==0.34.0",
    "httpx==0.28.1",
    "numpy==2.2.1",
    "scipy==1.15.0",
    "pandas==2.2.3",
    "scikit-learn==1.6.1",
    "pydantic==2.10.4",
    "pydantic-settings==2.7.1",
    "python-dotenv==1.0.1",
    "reportlab==4.1.0",
    "slowapi>=0.1.9",
    "google-genai>=1.0.0",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
]

[project.optional-dependencies]
bbn = [
    "pgmpy>=0.0.21",
]
sar = [
    "earthengine-api>=1.1.0",
]
dev = [
    "pytest==8.3.4",
    "pytest-cov==6.0.0",
    "pytest-asyncio==0.25.0",
    "httpx==0.28.1",
    "ruff==0.8.6",
    "mypy==1.14.1",
    "pandas-stubs==2.2.3.241126",
]

[tool.ruff]
target-version = "py310"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "RUF"]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.setuptools.packages.find]
where = ["."]
